+++ tests/core/contracts/test_extracting_event_data.py
@@ -9,6 +9,7 @@ from eth_utils.toolz import (
 )
 
 from web3.contract.utils import (
+    find_matching_event_abi,
     get_event_data,
 )
 from web3.exceptions import (
@@ -123,7 +124,8 @@ def test_event_data_extraction(
     assert len(txn_receipt["logs"]) == 1
     log_entry = txn_receipt["logs"][0]
 
-    event_abi = emitter._find_matching_event_abi(event_name)
+    contract_abi = emitter.abi
+    event_abi = find_matching_event_abi(contract_abi, event_name)
 
     event_topic = getattr(emitter_contract_log_topics, event_name)
     is_anonymous = event_abi["anonymous"]
@@ -178,7 +180,8 @@ def test_event_data_extraction_bytes(
     log_entry = txn_receipt["logs"][0]
 
     event_name = "LogListArgs"
-    event_abi = emitter._find_matching_event_abi(event_name)
+    contract_abi = emitter.abi
+    event_abi = find_matching_event_abi(contract_abi, event_name)
 
     event_topic = getattr(emitter_contract_log_topics, event_name)
 
@@ -229,7 +232,8 @@ def test_event_data_extraction_bytes_non_strict(
     log_entry = txn_receipt["logs"][0]
 
     event_name = "LogListArgs"
-    event_abi = non_strict_emitter._find_matching_event_abi(event_name)
+    contract_abi = non_strict_emitter.abi
+    event_abi = find_matching_event_abi(contract_abi, event_name)
 
     event_topic = getattr(emitter_contract_log_topics, event_name)
 
@@ -274,7 +278,9 @@ def test_dynamic_length_argument_extraction(
     assert len(txn_receipt["logs"]) == 1
     log_entry = txn_receipt["logs"][0]
 
-    event_abi = emitter._find_matching_event_abi("LogDynamicArgs")
+    event_name = "LogDynamicArgs"
+    contract_abi = emitter.abi
+    event_abi = find_matching_event_abi(contract_abi, event_name)
 
     event_topic = emitter_contract_log_topics.LogDynamicArgs
     assert event_topic in log_entry["topics"]
@@ -294,7 +300,7 @@ def test_dynamic_length_argument_extraction(
     assert event_data["blockNumber"] == txn_receipt["blockNumber"]
     assert event_data["transactionIndex"] == txn_receipt["transactionIndex"]
     assert is_same_address(event_data["address"], emitter.address)
-    assert event_data["event"] == "LogDynamicArgs"
+    assert event_data["event"] == event_name
 
 
 def test_argument_extraction_strict_bytes_types(
@@ -309,7 +315,9 @@ def test_argument_extraction_strict_bytes_types(
     log_entry = txn_receipt["logs"][0]
     assert len(log_entry["topics"]) == 2
 
-    event_abi = emitter._find_matching_event_abi("LogListArgs")
+    event_name = "LogListArgs"
+    contract_abi = emitter.abi
+    event_abi = find_matching_event_abi(contract_abi, event_name)
 
     event_topic = emitter_contract_log_topics.LogListArgs
     assert event_topic in log_entry["topics"]
