+++ tests/core/utilities/test_construct_event_filter_params.py
@@ -1,6 +1,6 @@
 import pytest
 
-from web3._utils.filters import (
+from web3.contract.utils import (
     construct_event_filter_params,
 )
